+++ test/unit/lib/configuration/variables/sources/file.test.js
@@ -22,8 +22,17 @@ describe('test/unit/lib/configuration/variables/sources/file.test.js', () => {
       jsFunction: '${file(file-function.js)}',
       jsPropertyFunction: '${file(file-property-function.js):property}',
       addressSupport: '${file(file.json):result}',
-      nonExistingYaml: '${file(not-existing.yaml)}',
-      nonExistingJs: '${file(not-existing.js)}',
+      nonExistingYaml: '${file(not-existing.yaml), null}',
+      nonExistingJson: '${file(not-existing.json), null}',
+      nonExistingJs: '${file(not-existing.js), null}',
+      primitiveAddress: '${file(file-primitive.json):someProperty}',
+      jsFilePromiseRejected: '${file(file-promise-rejected.js)}',
+      jsFilePromiseRejectedNonError: '${file(file-promise-rejected-non-error.js)}',
+      jsFileFunctionErrored: '${file(file-function-errored.js)}',
+      jsFileFunctionErroredNonError: '${file(file-function-errored-non-error.js)}',
+      jsFilePropertyFunctionErrored: '${file(file-property-function-errored.js):property}',
+      jsFilePropertyFunctionErroredNonError:
+        '${file(file-property-function-errored-non-error.js):property}',
       notFile: '${file(dir.yaml)}',
       noParams: '${file:}',
       noParams2: '${file():}',
@@ -32,7 +41,7 @@ describe('test/unit/lib/configuration/variables/sources/file.test.js', () => {
       invalidJson: '${file(invalid.json)}',
       invalidJs: '${file(invalid.js)}',
       invalidJs2: '${file(invalid2.js)}',
-      invalidExt: '${file(invalid.ext)}',
+      nonStandardExt: '${file(non-standard.ext)}',
     };
     variablesMeta = resolveMeta(configuration);
     await resolve({
@@ -68,11 +77,52 @@ describe('test/unit/lib/configuration/variables/sources/file.test.js', () => {
   it('should support "address" argument', () =>
     expect(configuration.addressSupport).to.equal('json'));
 
-  it('should report with an error non existing files', () =>
-    expect(variablesMeta.get('nonExistingYaml').error.code).to.equal('VARIABLE_RESOLUTION_ERROR'));
+  it('should report with null non existing files', () =>
+    expect(configuration.nonExistingYaml).to.equal(null));
 
-  it('should report with an error non existing JS files', () =>
-    expect(variablesMeta.get('nonExistingJs').error.code).to.equal('VARIABLE_RESOLUTION_ERROR'));
+  it('should report with null non existing JSON files', () =>
+    expect(configuration.nonExistingJson).to.equal(null));
+
+  it('should report with null non existing JS files', () =>
+    expect(configuration.nonExistingJs).to.equal(null));
+
+  it('should resolve plain text content on unrecognized extension', () =>
+    // .trim() as depending on local .git settings and OS (Windows or other)
+    // checked out fixture may end with differen type of EOL (\n on linux, and \r\n on Windows)
+    expect(configuration.nonStandardExt.trim()).to.equal('result: non-standard'.trim()));
+
+  it('should report with an error address argument on primitive content', () =>
+    expect(variablesMeta.get('primitiveAddress').error.code).to.equal('VARIABLE_RESOLUTION_ERROR'));
+
+  it('should report with an error promise rejected with error', () =>
+    expect(variablesMeta.get('jsFilePromiseRejected').error.code).to.equal(
+      'VARIABLE_RESOLUTION_ERROR'
+    ));
+
+  it('should report with an error promise rejected with non error value', () =>
+    expect(variablesMeta.get('jsFilePromiseRejectedNonError').error.code).to.equal(
+      'VARIABLE_RESOLUTION_ERROR'
+    ));
+
+  it('should report with an error function resolver that crashes with error', () =>
+    expect(variablesMeta.get('jsFileFunctionErrored').error.code).to.equal(
+      'VARIABLE_RESOLUTION_ERROR'
+    ));
+
+  it('should report with an error function resolver that crashes not with error', () =>
+    expect(variablesMeta.get('jsFileFunctionErroredNonError').error.code).to.equal(
+      'VARIABLE_RESOLUTION_ERROR'
+    ));
+
+  it('should report with an error property function resolver that crashes with error', () =>
+    expect(variablesMeta.get('jsFilePropertyFunctionErrored').error.code).to.equal(
+      'VARIABLE_RESOLUTION_ERROR'
+    ));
+
+  it('should report with an error property function resolver that crashes not with error', () =>
+    expect(variablesMeta.get('jsFilePropertyFunctionErroredNonError').error.code).to.equal(
+      'VARIABLE_RESOLUTION_ERROR'
+    ));
 
   it('should report with an error non file paths', () =>
     expect(variablesMeta.get('notFile').error.code).to.equal('VARIABLE_RESOLUTION_ERROR'));
@@ -95,9 +145,6 @@ describe('test/unit/lib/configuration/variables/sources/file.test.js', () => {
     expect(variablesMeta.get('invalidJs').error.code).to.equal('VARIABLE_RESOLUTION_ERROR');
     expect(variablesMeta.get('invalidJs2').error.code).to.equal('VARIABLE_RESOLUTION_ERROR');
   });
-
-  it('should report with an error an unrecognized extension', () =>
-    expect(variablesMeta.get('invalidExt').error.code).to.equal('VARIABLE_RESOLUTION_ERROR'));
 
   it('should not support function resolvers in "js" file sources not confirmed to work with new resolver', async () => {
     configuration = {